% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arguments}
\alias{check_arguments}
\alias{check_task_type}
\alias{check_features}
\alias{check_tmt}
\alias{check_weight}
\alias{check_y}
\alias{check_block}
\alias{check_cluster}
\alias{check_block_cluster}
\alias{check_method}
\alias{check_test_split}
\alias{check_tune_split}
\alias{check_learner}
\alias{check_loss}
\alias{check_permutations}
\alias{check_parallel}
\title{Check Arguments}
\usage{
check_task_type(task_type)

check_features(features)

check_tmt(tmt, task_type)

check_weight(weight)

check_y(y)

check_block(block)

check_cluster(cluster)

check_block_cluster(block, cluster)

check_method(test_method)

check_test_split(test_split, test_method)

check_tune_split(tune_split)

check_learner(learner)

check_loss(loss, task_type)

check_permutations(permutations, task_type)

check_parallel(parallel)
}
\arguments{
\item{task_type}{"binary" | "regression" | "multiclass" \cr
The kinds of models and predictions to make for treatment status.
"binary" for binary treatment, "regression" for continuous
treatment, and "multiclass" for discrete multilevel treatment.}

\item{features}{matrix | (named) list of matrices \cr
Features to use when predicting treatment. Can be a matrix or a (named) list
of matrices.}

\item{tmt}{vector \cr
A vector of treatment assignments.}

\item{weight}{numeric vector, optional \cr
Observation weights.}

\item{y}{matrix, optional
Outcomes of the experiment. If provided, these will be used to estimate
treatment effects. If you want to use the outcomes to predict treatment
effect, they need to be included in \code{features}. NULL skips treatment
effect estimation.}

\item{block}{vector, optional \cr
A vector of block IDs, if treatment was randomly assigned within blocks.}

\item{cluster}{vector, optional
A vector of cluster IDs, if treatment was randomly assigned to clusters
of observations. All observations in the same cluster must have been 
assigned the same treatment.}

\item{test_method}{"cv" | "holdout" \cr
Whether to use cross-validation ("cv") or a holdout set ("holdout") for
hypothesis tests. Using a holdout usually reduces the statistical power of
the procedure, but will also usually be faster.}

\item{test_split}{proportion | integer | integer vector \cr
Only used when test_method = "holdout". Either the proportion of observations
to place in the holdout set, or an integer vector that assigns observations
to the holdout set. If given a vector, entries with NAs will be excluded
from both the training and holdout sets..}

\item{tune_split}{integer | integer vector \cr
Either the number of folds for tuning, or an integer vector that assigns
observations to tuning folds. Entries with NAs will not be used for tuning.}

\item{learner}{object of class "learner" or "constructor" \cr
A predictive model.}

\item{loss}{function or (named) list of functions \cr
A function to measure how good predictions are. Lower values indicate
better predictions. If provided as a list of functions, each function will
be used to evaluate predictions separately.}

\item{permutations}{integer | list of permuted tmt \cr
Either the number of permutations to perform or a list of permuted treatment
assignments in the holdout set. NULL skips the permutation test.}

\item{parallel}{character vector \cr
Steps of analysis to parallelize. Can be any combination of "tune",
"train", "predict", "perm_test", and "tmt_hat_struct". Only effective when
comparing a list of x or a list of learners.}
}
\description{
Utilities to checks argument for compatibility with this
package's expectations for inputs. They will throw errors if expectations
are violated.
}
