% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_tbs}
\alias{combine_tbs}
\title{Rowbind and Index a List of Tables}
\usage{
combine_tbs(list_of_tb, list_names, new_column)
}
\arguments{
\item{list_of_tb}{list \cr
A list of data.frames}

\item{list_names}{character \cr
A character vector naming each table in list_of_tb}

\item{new_column}{character(1) \cr
Name of the index column to create in the result. Will hold entries from
list_names.}
}
\value{
data.frame
}
\description{
Rowbinds a list of tables and creates a new indexing column
to identify which table each row in the result came from.
}
