% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_funs.R
\docType{data}
\name{loss_funs}
\alias{loss_funs}
\title{Common Loss Functions}
\format{An object of class \code{list} of length 5.}
\usage{
loss_funs
}
\description{
Common Loss Functions
}
\details{
loss functions take three arguments: (actual, predicted, weight),
and return a loss. Current loss functions include
\itemize{
  \item rmse: root-mean-squared error
  \item mse: mean-squared error
  \item roc_area_above: the area above the receiver operating curve. This is
  the empirical probability that a randomly selected control unit receives
  a higher predicted treatment status than a randomly selected treatment
  unit.
  \item neg_cov: the negative covariance between predicted and observed
  treatment status.
  \item neg_bern_loglik: the negative Bernoulli log-likelihood
}
}
\keyword{datasets}
