% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds.R
\name{make_folds}
\alias{make_folds}
\title{Divide Data into Tuning and Training Sets}
\usage{
make_folds(test_method, tune_split, test_split, task_type, tmt,
  block = NULL, cluster = NULL)
}
\arguments{
\item{test_method}{"cv" | "holdout" \cr
Whether to use cross-validation ("cv") or a holdout set ("holdout") for
hypothesis tests. Using a holdout usually reduces the statistical power of
the procedure, but will also usually be faster.}

\item{tune_split}{integer | integer vector \cr
Either the number of folds for tuning, or an integer vector that assigns
observations to tuning folds. Entries with NAs will not be used for tuning.}

\item{test_split}{proportion | integer | integer vector \cr
Only used when test_method = "holdout". Either the proportion of observations
to place in the holdout set, or an integer vector that assigns observations
to the holdout set. If given a vector, entries with NAs will be excluded
from both the training and holdout sets..}

\item{task_type}{"binary" | "regression" | "multiclass" \cr
The kinds of models and predictions to make for treatment status.
"binary" for binary treatment, "regression" for continuous
treatment, and "multiclass" for discrete multilevel treatment.}

\item{tmt}{vector \cr
A vector of treatment assignments.}

\item{block}{vector, optional \cr
A vector of block IDs, if treatment was randomly assigned within blocks.}

\item{cluster}{vector, optional
A vector of cluster IDs, if treatment was randomly assigned to clusters
of observations. All observations in the same cluster must have been 
assigned the same treatment.}
}
\description{
Divide Data into Tuning and Training Sets
}
