% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perm_p_val}
\alias{perm_p_val}
\title{Calculate Permutation P-Value}
\usage{
perm_p_val(observed, permuted)
}
\arguments{
\item{observed}{numeric \cr
Observed loss.}

\item{permuted}{numeric \cr
Permutation losses.}
}
\description{
Calculate Permutation P-Value
}
