% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{perm_test}
\alias{perm_test}
\title{Perform a Permutation Test}
\usage{
perm_test(tmt, pred, permuted_tmt = NULL, permuted_pred = NULL,
  weight = NULL, loss = list(rmse = loss_funs$rmse))
}
\arguments{
\item{tmt}{vector \cr
A vector of treatment assignments.}

\item{pred}{numeric | matrix  \cr
a) For purely binary classification losses: The predicted probabilities for 
the positive class as a numeric vector. b) For multiclass classification 
losses: The predicted probabilities for all classes, always as a 
numeric matrix, where columns are named with class labels.}

\item{permuted_tmt}{list \cr
A list of permuted treatment indicators}

\item{permuted_pred}{NULL | list \cr
A list of permuted predictions}

\item{weight}{numeric vector, optional \cr
Observation weights.}

\item{loss}{function or (named) list of functions \cr
A function to measure how good predictions are. Lower values indicate
better predictions. If provided as a list of functions, each function will
be used to evaluate predictions separately.}
}
\description{
Perform a Permutation Test
}
\details{
If the observed tmt vector is included in the permuted_tmt list,
then the p-value returned is mean(permuted_loss <= observed_loss). But, if
the observed tmt vector is not in the permuted_tmt list, the p-value
returned is mean(c(observed_loss, permuted_loss) <= observed_loss). This
ensures that the estimated p-value is never exactly 0.
}
