% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_tmt.R
\name{permute_tmt}
\alias{permute_tmt}
\title{Permute a Treatment Assignment Vector}
\usage{
permute_tmt(tmt, n_perm = 1L, block = NULL, cluster = NULL,
  max_tries = 10)
}
\arguments{
\item{tmt}{vector \cr
A vector of treatment assignments.}

\item{n_perm}{integer(1) \cr
The number of permutations to generate.}

\item{block}{vector \cr
Optional, observation-level block IDs.}

\item{cluster}{vector \cr
Optional, observation-level cluster IDs.}

\item{max_tries}{integer(1) \cr
The number of attempts to make at producing `n_perm` unique permutations.
After `max_tries`, the unique permutations generated so far will be returned.}
}
\value{
[vector] Treatment assignments, permuted by cluster and within block.
Removes duplicates, returning only unique permutations
}
\description{
Generate unique permutations of treatment assignment.
}
