% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_tmt.R
\name{resample_treatment_clusterwise}
\alias{resample_treatment_clusterwise}
\title{Permute Treatment Assignment at Cluster Level}
\usage{
resample_treatment_clusterwise(cluster_tmt, cluster_block,
  obs_cluster_index)
}
\arguments{
\item{cluster_tmt}{Treatment assigned to each cluster.}

\item{cluster_block}{Block each cluster belongs to.}

\item{obs_cluster_index}{Cluster each observation belongs to.}
}
\description{
Permute Treatment Assignment at Cluster Level
}
