% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_tmt_hat_struct.R
\name{summarize_tmt_hat_struct}
\alias{summarize_tmt_hat_struct}
\title{Estimate Treatment Effect Given Outcomes and Treatment Predictions}
\usage{
summarize_tmt_hat_struct(tmt_hat, y, tmt, subset, task_type)
}
\arguments{
\item{tmt_hat}{vector | matrix \cr
Estimated treatment assignment. Probability of treatment assignment for
binary tasks, dosage for continuous tasks, and matrix of control and
treatment probabilities for multiclass tasks (one probability per column).}

\item{y}{matrix, optional
Outcomes of the experiment. If provided, these will be used to estimate
treatment effects. If you want to use the outcomes to predict treatment
effect, they need to be included in \code{features}. NULL skips treatment
effect estimation.}

\item{tmt}{vector \cr
A vector of treatment assignments.}

\item{subset}{vector \cr
Optional, a logical or integer vector identifying which observations to
use when estimating treatment effects.}

\item{task_type}{"binary" | "regression" | "multiclass" \cr
The kinds of models and predictions to make for treatment status.
"binary" for binary treatment, "regression" for continuous
treatment, and "multiclass" for discrete multilevel treatment.}
}
\description{
Estimate Treatment Effect Given Outcomes and Treatment Predictions
}
