% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_moments.R
\name{weighted_var}
\alias{weighted_var}
\alias{weighted_sd}
\title{Weighted Sampling Variance and SD}
\usage{
weighted_var(x, w, na.rm = FALSE)

weighted_sd(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{an object containing the values whose weighted variance or standard deviation
to calculate.}

\item{w}{a numerical vector of weights the same length as x giving the weights to use 
for elements of x.}

\item{na.rm}{a logical value indicating whether NA values in x should be stripped
before computaiton.}
}
\description{
Unbiased estimator of sampling variance, with weights.
}
